+++ photutils/segmentation/core.py
@@ -4,15 +4,121 @@ from __future__ import (absolute_import, division, print_function,
 from copy import deepcopy
 
 import numpy as np
-from astropy.utils import lazyproperty
+from astropy.utils import lazyproperty, deprecated
 
+from ..aperture import BoundingBox
 from ..utils.colormaps import random_cmap
 
 
-__all__ = ['SegmentationImage']
+__all__ = ['Segment', 'SegmentationImage']
 
 __doctest_requires__ = {('SegmentationImage', 'SegmentationImage.*'):
                         ['scipy', 'skimage']}
+
+
+class Segment(object):
+    """
+    Class for a single labeled region (segment) within a segmentation
+    image.
+
+    Parameters
+    ----------
+    segment_img : 2D int `~numpy.ndarray`
+        A 2D segmentation image where source regions are labeled by
+        different positive integer values.  A value of zero is reserved
+        for the background.
+
+    label : int
+        The segment label number.
+
+    slices : tuple of two slices
+        A tuple of two slices representing the minimal box that contains
+        the labeled region.
+
+    area : float
+        The area of the segment in pixels**2.
+    """
+
+    def __init__(self, segment_img, label, slices, area):
+        self._segment_img = segment_img
+        self.label = label
+        self.slices = slices
+        self.area = area
+
+    def __str__(self):
+        cls_name = '<{0}.{1}>'.format(self.__class__.__module__,
+                                      self.__class__.__name__)
+
+        cls_info = []
+        params = ['label', 'slices', 'bbox', 'area']
+        for param in params:
+            cls_info.append((param, getattr(self, param)))
+        fmt = ['{0}: {1}'.format(key, val) for key, val in cls_info]
+
+        return '{}\n'.format(cls_name) + '\n'.join(fmt)
+
+    def __repr__(self):
+        return self.__str__()
+
+    @lazyproperty
+    def data(self):
+        """
+        Cutout image of the segment, where pixels outside of the labeled
+        region are set to zero (i.e. neighboring segments within the
+        rectangular cutout image are not shown).
+        """
+
+        cutout = np.copy(self._segment_img[self.slices])
+        cutout[cutout != self.label] = 0
+
+        return cutout
+
+    def __array__(self):
+        """
+        Array representation of the labeled region (e.g., for
+        matplotlib).
+        """
+
+        return self.data
+
+    @lazyproperty
+    def bbox(self):
+        return BoundingBox(self.slices[1].start, self.slices[1].stop,
+                           self.slices[0].start, self.slices[0].stop)
+
+    def make_cutout(self, data, masked_array=False):
+        """
+        Create a (masked) cutout array from the input ``data`` using the
+        minimal bounding box of the segment (labeled region).
+
+        Parameters
+        ----------
+        data : 2D array-like
+            The data array from which to create the masked cutout array.
+            ``data`` must have the same shape as the segmentation image.
+
+        masked_array : bool, optional
+            If `True` then a `~numpy.ma.MaskedArray` will be created
+            where the mask is `True` for pixels outside of the segment
+            (labeled region).  If `False`, then a `~numpy.ndarray` will
+            be generated.
+
+        Returns
+        -------
+        result : 2D `~numpy.ndarray` or 2D `~numpy.ma.MaskedArray`
+            The 2D cutout array or masked array.
+        """
+
+        data = np.asanyarray(data)
+        if data.shape != self._segment_img.shape:
+            raise ValueError('data must have the same shape as the '
+                             'segmentation image.')
+
+        if masked_array:
+            mask = (self._segment_img[self.slices] != self.label)
+            return np.ma.masked_array(data[self.slices], mask=mask)
+        else:
+            return data[self.slices]
 
 
 class SegmentationImage(object):
@@ -22,14 +128,44 @@ class SegmentationImage(object):
     Parameters
     ----------
     data : array_like (int)
-        A 2D segmentation image where sources are labeled by different
-        positive integer values.  A value of zero is reserved for the
-        background.
+        A 2D segmentation image where source regions are labeled by
+        different positive integer values.  A value of zero is reserved
+        for the background.
     """
 
     def __init__(self, data):
-
         self.data = np.asanyarray(data, dtype=np.int)
+
+    def __getitem__(self, index):
+        return self.segments[index]
+
+    # python 2 only
+    def __getslice__(self, i, j):
+        return self.__getitem__(slice(i, j))
+
+    def __iter__(self):
+        for i in self.segments:
+            yield i
+
+    @lazyproperty
+    def segments(self):
+        """
+        A list of `Segment` objects.
+
+        The list starts with the *non-zero* label.  The returned list
+        has a length equal to the maximum label number.  If a label
+        number is missing from the segmentation image, then `None` is
+        returned instead of a `Segment` object.
+        """
+
+        segments = []
+        for i, slc in enumerate(self.slices):
+            if slc is None:
+                segments.append(None)
+            else:
+                segments.append(Segment(self.data, i+1, slc, self.areas[i]))
+
+        return segments
 
     @property
     def data(self):
@@ -44,18 +180,18 @@ class SegmentationImage(object):
         if np.min(value) < 0:
             raise ValueError('The segmentation image cannot contain '
                              'negative integers.')
+
+        if '_data' in self.__dict__:
+            # needed only when data is reassigned, not on init
+            self._reset_lazy_properties()
+
         self._data = value
-        # be sure to delete any lazy properties to reset their values.
-        del (self.data_masked, self.shape, self.labels, self.nlabels,
-             self.max, self.slices, self.areas, self.is_sequential)
 
-    @property
-    def array(self):
-        """
-        The 2D segmentation image.
-        """
-
-        return self._data
+    def _reset_lazy_properties(self):
+        """Reset all lazy properties."""
+        for key, value in self.__class__.__dict__.items():
+            if isinstance(value, lazyproperty):
+                self.__dict__.pop(key, None)
 
     def __array__(self):
         """
@@ -66,7 +202,12 @@ class SegmentationImage(object):
         return self._data
 
     @lazyproperty
+    @deprecated(0.5, alternative='data_ma')
     def data_masked(self):
+        return self.data_ma  # pragma: no cover
+
+    @lazyproperty
+    def data_ma(self):
         """
         A `~numpy.ma.MaskedArray` version of the segmentation image
         where the background (label = 0) has been masked.
@@ -83,7 +224,7 @@ class SegmentationImage(object):
         Parameters
         ----------
         data : array_like (int)
-            A 2D segmentation image where sources are labeled by
+            A 2D segmentation image where source regions are labeled by
             different positive integer values.  A value of zero is
             reserved for the background.
 
@@ -94,9 +235,9 @@ class SegmentationImage(object):
 
         Notes
         -----
-        This is a separate static method so it can be used on masked
-        versions of the segmentation image (cf.
-        ``~photutils.SegmentationImage.remove_masked_labels``.
+        This is a separate static method so it can be used in
+        :meth:`remove_masked_labels` on a masked version of the
+        segmentation image.
 
         Examples
         --------
@@ -111,6 +252,7 @@ class SegmentationImage(object):
         array([1, 3, 4, 5, 7])
         """
 
+        # np.unique also sorts elements
         return np.unique(data[data != 0])
 
     @lazyproperty
@@ -134,31 +276,54 @@ class SegmentationImage(object):
         return len(self.labels)
 
     @lazyproperty
+    @deprecated(0.5, alternative='max_label')
     def max(self):
+        return self.max_label  # pragma: no cover
+
+    @lazyproperty
+    def max_label(self):
         """The maximum non-zero label in the segmentation image."""
 
-        return np.max(self.data)
+        return np.max(self.labels)
 
     @lazyproperty
     def slices(self):
-        """The minimal bounding box slices for each labeled region."""
+        """
+        A list of tuples, where each tuple contains two slices
+        representing the minimal box that contains the labeled region.
+
+        The list starts with the *non-zero* label.  The returned list
+        has a length equal to the maximum label number.  If a label
+        number is missing, then `None` is returned for that list element
+        instead of a slice.
+        """
 
         from scipy.ndimage import find_objects
+
         return find_objects(self._data)
 
     @lazyproperty
     def areas(self):
-        """The areas (in pixel**2) of all labeled regions."""
+        """
+        A 1D array of areas (in pixel**2) of the non-zero labeled
+        regions.
 
-        return np.bincount(self.data.ravel())
+        The `~numpy.ndarray` starts with the *non-zero* label The
+        returned array has a length equal to the maximum label number.
+        If a label number is missing, then 0 is returned for that array
+        element.
+        """
 
-    def area(self, labels):
+        return np.bincount(self.data.ravel())[1:]
+
+    @deprecated(0.5, alternative='areas[labels-1]')
+    def area(self, labels):  # pragma: no cover
         """
         The areas (in pixel**2) of the regions for the input labels.
 
         Parameters
         ----------
-        labels : int, array-like (1D, int)
+        labels : int, 1D array-like (int)
             The label(s) for which to return areas.
 
         Returns
@@ -168,15 +333,20 @@ class SegmentationImage(object):
         """
 
         labels = np.atleast_1d(labels)
-        for label in labels:
-            self.check_label(label, allow_zero=True)
-        return self.areas[labels]
+        self.check_labels(labels)
+
+        return self.areas[labels - 1]
 
     @lazyproperty
+    @deprecated(0.5, alternative='is_consecutive')
     def is_sequential(self):
+        return self.is_consecutive  # pragma: no cover
+
+    @lazyproperty
+    def is_consecutive(self):
         """
         Determine whether or not the non-zero labels in the segmenation
-        image are sequential (with no missing values).
+        image are consecutive (i.e. no missing values).
         """
 
         if (self.labels[-1] - self.labels[0] + 1) == self.nlabels:
@@ -184,16 +354,57 @@ class SegmentationImage(object):
         else:
             return False
 
+    @lazyproperty
+    def missing_labels(self):
+        """
+        A 1D `~numpy.ndarray` of the sorted non-zero labels that are
+        missing in the consecutive sequence from zero to the maximum
+        label number.
+        """
+
+        return np.array(sorted(set(range(0, self.max_label + 1)).
+                               difference(np.insert(self.labels, 0, 0))))
+
     def copy(self):
         """
         Return a deep copy of this class instance.
-
-        Deep copy is used so that all attributes and values are copied.
         """
 
         return deepcopy(self)
 
-    def check_label(self, label, allow_zero=False):
+    def check_labels(self, labels):
+        """
+        Check that the input label(s) are valid label numbers within the
+        segmentation image.
+
+        Parameters
+        ----------
+        labels : int, 1D array-like (int)
+            The label(s) to check.
+
+        Raises
+        ------
+        ValueError
+            If any input ``labels`` are invalid.
+        """
+
+        labels = np.atleast_1d(labels)
+        bad_labels = set()
+
+        # check for positive label numbers
+        idx = np.where(labels <= 0)[0]
+        if len(idx) > 0:
+            bad_labels.update(labels[idx])
+
+        # check if label is in the segmentation image
+        bad_labels.update(np.setdiff1d(labels, self.labels))
+
+        bad_labels = tuple(bad_labels)
+        if len(bad_labels) > 0:
+            raise ValueError('labels {} are invalid'.format(bad_labels))
+
+    @deprecated(0.5, alternative='check_labels')
+    def check_label(self, label, allow_zero=False):  # pragma: no cover
         """
         Check for a valid label label number within the segmentation
         image.
@@ -248,7 +459,7 @@ class SegmentationImage(object):
 
         from matplotlib import colors
 
-        cmap = random_cmap(self.max + 1, random_state=random_state)
+        cmap = random_cmap(self.max_label + 1, random_state=random_state)
 
         if background_color is not None:
             cmap.colors[0] = colors.hex2color(background_color)
@@ -344,9 +555,13 @@ class SegmentationImage(object):
             data[np.where(data == label)] = new_label
             self.data = data     # needed to call the data setter
 
+    @deprecated(0.5, alternative='relabel_consecutive()')
     def relabel_sequential(self, start_label=1):
+        return self.relabel_consecutive(start_label=start_label)  # pragma: no cover
+
+    def relabel_consecutive(self, start_label=1):
         """
-        Relabel the label numbers sequentially, such that there are no
+        Relabel the label numbers consecutively, such that there are no
         missing label numbers (up to the maximum label number).
 
         Parameters
@@ -364,7 +579,7 @@ class SegmentationImage(object):
         ...                           [7, 0, 0, 0, 0, 5],
         ...                           [7, 7, 0, 5, 5, 5],
         ...                           [7, 7, 0, 0, 5, 5]])
-        >>> segm.relabel_sequential()
+        >>> segm.relabel_consecutive()
         >>> segm.data
         array([[1, 1, 0, 0, 3, 3],
                [0, 0, 0, 0, 0, 3],
@@ -377,10 +592,10 @@ class SegmentationImage(object):
         if start_label <= 0:
             raise ValueError('start_label must be > 0.')
 
-        if self.is_sequential and (self.labels[0] == start_label):
+        if self.is_consecutive and (self.labels[0] == start_label):
             return
 
-        forward_map = np.zeros(self.max + 1, dtype=np.int)
+        forward_map = np.zeros(self.max_label + 1, dtype=np.int)
         forward_map[self.labels] = np.arange(self.nlabels) + start_label
         self.data = forward_map[self.data]
 
@@ -396,7 +611,7 @@ class SegmentationImage(object):
 
         relabel : bool, optional
             If `True`, then the segmentation image will be relabeled
-            such that the labels are in sequential order starting from
+            such that the labels are in consecutive order starting from
             1.
 
         Examples
@@ -449,7 +664,7 @@ class SegmentationImage(object):
 
         relabel : bool, optional
             If `True`, then the segmentation image will be relabeled
-            such that the labels are in sequential order starting from
+            such that the labels are in consecutive order starting from
             1.
 
         Examples
@@ -488,7 +703,7 @@ class SegmentationImage(object):
 
         self.relabel(labels, new_label=0)
         if relabel:
-            self.relabel_sequential()
+            self.relabel_consecutive()
 
     def remove_border_labels(self, border_width, partial_overlap=True,
                              relabel=False):
@@ -510,7 +725,7 @@ class SegmentationImage(object):
 
         relabel : bool, optional
             If `True`, then the segmentation image will be relabeled
-            such that the labels are in sequential order starting from
+            such that the labels are in consecutive order starting from
             1.
 
         Examples
@@ -572,14 +787,14 @@ class SegmentationImage(object):
             pixels.
 
         partial_overlap : bool, optional
-            If this is set to `True` (the default), a segment that
-            partially extends into a masked region will also be removed.
-            Segments that are completely within a masked region are
-            always removed.
+            If this is set to `True` (default), a segment that partially
+            extends into a masked region will also be removed.  Segments
+            that are completely within a masked region are always
+            removed.
 
         relabel : bool, optional
             If `True`, then the segmentation image will be relabeled
-            such that the labels are in sequential order starting from
+            such that the labels are in consecutive order starting from
             1.
 
         Examples